refactor(shapesMap): moved out shapesMap to seperate file * removed switch case * work in progress
