+++ packages/svg-deserializer/index-csg.js
@@ -484,41 +484,6 @@ const groupValue = function (name) {
   }
   return null
 }
-
-let typesMap = {
-  group: (obj, onCAG) => {
-    const subCag = codify(obj)
-    onCAG = onCAG.union(subCag)
-    return onCAG
-  },
-  rectangle: (obj, svgUnitsPmm, svgUnitsX, svgUnitsY) => {
-    const x = cagLengthX(obj.x, svgUnitsPmm, svgUnitsX)
-    const y = (0 - cagLengthY(obj.y, svgUnitsPmm, svgUnitsY))
-    const w = cagLengthX(obj.width, svgUnitsPmm, svgUnitsX)
-    const h = cagLengthY(obj.height, svgUnitsPmm, svgUnitsY)
-    const rx = cagLengthX(obj.rx, svgUnitsPmm, svgUnitsX)
-    const ry = cagLengthY(obj.ry, svgUnitsPmm, svgUnitsY)
-    if (w > 0 && h > 0) {
-      x = (x + (w / 2)).toFixed(4)  // position the object via the center
-      y = (y - (h / 2)).toFixed(4)  // position the object via the center
-      if (rx === 0) {
-        return CAG.rectangle({center: [x, y], radius: [w / 2, h / 2]})
-      } else {
-        return CAG.roundedRectangle({center: [x, y], radius: [w / 2, h / 2], roundradius: rx})
-      }
-    }
-  },
-  circle: (obj, svgUnitsPmm, svgUnitsX, svgUnitsY, svgUnitsV) => {
-    const x = cagLengthX(obj.x, svgUnitsPmm, svgUnitsX)
-    const y = (0 - cagLengthY(obj.y, svgUnitsPmm, svgUnitsY))
-    const r = cagLengthP(obj.radius, svgUnitsPmm, svgUnitsV)
-    if (r > 0) {
-      return CAG.circle({center: [x, y], radius: [r]})
-    }
-  }
-
-}
-
 const codify = function (group) {
   const level = svgGroups.length
   // add this group to the heiarchy
@@ -540,380 +505,15 @@ const codify = function (group) {
   for (i = 0; i < group.objects.length; i++) {
     console.log('bla', group.objects)
     var obj = group.objects[i]
-    const onCAG = new CAG() //
-    switch (obj.type) {
-      case 'group':
-        //cag from nested element
-        const subCag = codify(obj)
-        onCAG = onCAG.union(subCag)
-        break
-      case 'rect':
-        var x = cagLengthX(obj.x, svgUnitsPmm, svgUnitsX)
-        var y = (0 - cagLengthY(obj.y, svgUnitsPmm, svgUnitsY))
-        var w = cagLengthX(obj.width, svgUnitsPmm, svgUnitsX)
-        var h = cagLengthY(obj.height, svgUnitsPmm, svgUnitsY)
-        var rx = cagLengthX(obj.rx, svgUnitsPmm, svgUnitsX)
-        var ry = cagLengthY(obj.ry, svgUnitsPmm, svgUnitsY)
-        if (w > 0 && h > 0) {
-          x = (x + (w / 2)).toFixed(4)  // position the object via the center
-          y = (y - (h / 2)).toFixed(4)  // position the object via the center
-          if (rx === 0) {
-            objects.push(CAG.rectangle({center: [x, y], radius: [w / 2, h / 2]}))
-          } else {
-            objects.push(CAG.roundedRectangle({center: [x, y], radius: [w / 2, h / 2], roundradius: rx}))
-          }
-        }
-        break
-      case 'circle':
-        var x = cagLengthX(obj.x, svgUnitsPmm, svgUnitsX)
-        var y = (0 - cagLengthY(obj.y, svgUnitsPmm, svgUnitsY))
-        var r = cagLengthP(obj.radius, svgUnitsPmm, svgUnitsV)
-        if (r > 0) {
-          objects.push(CAG.circle({center: [x, y], radius: [r]}))
-        }
-        break
-      case 'ellipse':
-        var rx = cagLengthX(obj.rx, svgUnitsPmm, svgUnitsX)
-        var ry = cagLengthY(obj.ry, svgUnitsPmm, svgUnitsY)
-        var cx = cagLengthX(obj.cx, svgUnitsPmm, svgUnitsX)
-        var cy = (0 - cagLengthY(obj.cy, svgUnitsPmm, svgUnitsY))
-        if (rx > 0 && ry > 0) {
-          objects.push(CAG.ellipse({center: [cx, cy], radius: [rx, ry]}))
-        }
-        break
-      case 'line':
-        var x1 = cagLengthX(obj.x1, svgUnitsPmm, svgUnitsX)
-        var y1 = (0 - cagLengthY(obj.y1), svgUnitsPmm, svgUnitsY)
-        var x2 = cagLengthX(obj.x2, svgUnitsPmm, svgUnitsX)
-        var y2 = (0 - cagLengthY(obj.y2, svgUnitsPmm, svgUnitsY))
-        var r = cssPxUnit // default
-        if ('strokeWidth' in obj) {
-          r = cagLengthP(obj.strokeWidth, svgUnitsPmm, svgUnitsV) / 2
-        } else {
-          var v = groupValue('strokeWidth')
-          if (v !== null) {
-            r = cagLengthP(v, svgUnitsPmm, svgUnitsV) / 2
-          }
-        }
-        const tmpObj = new CSG.Path2D([[x1, y1], [x2, y2]], false)
-          .expandToCAG(r, CSG.defaultResolution2D)
-        objects.push(tmpObj)
+    const onCAG = require('./shapesMapCsg')(codify, svgUnitsPmm, svgUnitsX, svgUnitsY, svgUnitsV)
 
-        break
-      case 'polygon':
-        {
-          let points = []
-          var j = 0
-          for (j = 0; j < obj.points.length; j++) {
-            var p = obj.points[j]
-            if ('x' in p && 'y' in p) {
-              var x = cagLengthX(p.x, svgUnitsPmm, svgUnitsX)
-              var y = (0 - cagLengthY(p.y, svgUnitsPmm, svgUnitsY))
-              points.push([x, y])
-            }
-          }
-          let tmpObj = new CSG.Path2D(points, true).innerToCAG()
-          objects.push(tmpObj)
-        }
-        break
-      case 'polyline': {
-        let points = []
-        var r = cssPxUnit // default
-        if ('strokeWidth' in obj) {
-          r = cagLengthP(obj.strokeWidth, svgUnitsPmm, svgUnitsV) / 2
-        } else {
-          var v = groupValue('strokeWidth')
-          if (v !== null) {
-            r = cagLengthP(v, svgUnitsPmm, svgUnitsV) / 2
-          }
-        }
-        var j = 0
-        for (j = 0; j < obj.points.length; j++) {
-          var p = obj.points[j]
-          if ('x' in p && 'y' in p) {
-            var x = cagLengthX(p.x, svgUnitsPmm, svgUnitsX)
-            var y = (0 - cagLengthY(p.y, svgUnitsPmm, svgUnitsY))
-            points.push([x, y])
-          }
-        }
-        let tmpObj = new CSG.Path2D(points, false).expandToCAG(r, CSG.defaultResolution2D)
-        objects.push(tmpObj)
-      }
-        break
-      case 'path':
-        var pathCag = new CAG()
-        let currentPath
-        let paths = {}
-
-        var r = cssPxUnit // default
-        if ('strokeWidth' in obj) {
-          r = cagLengthP(obj.strokeWidth, svgUnitsPmm, svgUnitsV) / 2
-        } else {
-          var v = groupValue('strokeWidth')
-          if (v !== null) {
-            r = cagLengthP(v, svgUnitsPmm, svgUnitsV) / 2
-          }
-        }
-      // Note: All values are SVG values
-        var sx = 0     // starting position
-        var sy = 0
-        var cx = 0     // current position
-        var cy = 0
-        var pi = 0     // current path index
-        var pn = on + pi // current path name
-        var pc = false // current path closed
-        var bx = 0     // 2nd control point from previous C command
-        var by = 0     // 2nd control point from previous C command
-        var qx = 0     // 2nd control point from previous Q command
-        var qy = 0     // 2nd control point from previous Q command
-        var j = 0
-        for (j = 0; j < obj.commands.length; j++) {
-          var co = obj.commands[j]
-          var pts = co.p
-          // console.log('postion: ['+cx+','+cy+'] before '+co.c);
-          switch (co.c) {
-            case 'm': // relative move to X,Y
-            // special case, if at beginning of path then treat like absolute M
-              if (j === 0) {
-                cx = 0; cy = 0
-              }
-            // close the previous path
-              if (pi > 0 && pc === false) {
-                paths[pn].expandToCAG(CSG.defaultResolution2D)
-                // code += indent + pn + ' = ' + pn + '.expandToCAG(' + r + ',CSG.defaultResolution2D);\n'
-              }
-            // open a new path
-              if (pts.length >= 2) {
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                pi++
-                pn = on + pi
-                pc = false
-                paths[pn] = new CSG.Path2D([[svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]], false)
-                sx = cx; sy = cy
-              }
-            // optional implicit relative lineTo (cf SVG spec 8.3.2)
-              while (pts.length >= 2) {
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'M': // absolute move to X,Y
-            // close the previous path
-              if (pi > 0 && pc === false) {
-                paths[pn] = paths[pn].expandToCAG(CSG.defaultResolution2D)
-              }
-            // open a new path
-              if (pts.length >= 2) {
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                pi++
-                pn = on + pi
-                pc = false
-                paths[pn] = new CSG.Path2D([[svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]], false)
-                sx = cx; sy = cy
-              }
-            // optional implicit absolute lineTo (cf SVG spec 8.3.2)
-              while (pts.length >= 2) {
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'a': // relative elliptical arc
-              while (pts.length >= 7) {
-                var rx = parseFloat(pts.shift())
-                var ry = parseFloat(pts.shift())
-                var ro = 0 - parseFloat(pts.shift())
-                var lf = (pts.shift() === '1')
-                var sf = (pts.shift() === '1')
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn].appendArc([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)], {xradius: svg2cagX(rx), yradius: svg2cagY(ry), xaxisrotation: ro, clockwise: sf, large: lf})
-              }
-              break
-            case 'A': // absolute elliptical arc
-              while (pts.length >= 7) {
-                var rx = parseFloat(pts.shift())
-                var ry = parseFloat(pts.shift())
-                var ro = 0 - parseFloat(pts.shift())
-                var lf = (pts.shift() === '1')
-                var sf = (pts.shift() === '1')
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn].appendArc([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)], {xradius: svg2cagX(rx), yradius: svg2cagY(ry), xaxisrotation: ro, clockwise: sf, large: lf})
-              }
-              break
-            case 'c': // relative cubic Bézier
-              while (pts.length >= 6) {
-                var x1 = cx + parseFloat(pts.shift())
-                var y1 = cy + parseFloat(pts.shift())
-                bx = cx + parseFloat(pts.shift())
-                by = cy + parseFloat(pts.shift())
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(x1), svg2cagY(y1)], [svg2cagX(bx), svg2cagY(by)], [svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(bx, by, cx, cy)
-                bx = rf[0]
-                by = rf[1]
-              }
-              break
-            case 'C': // absolute cubic Bézier
-              while (pts.length >= 6) {
-                var x1 = parseFloat(pts.shift())
-                var y1 = parseFloat(pts.shift())
-                bx = parseFloat(pts.shift())
-                by = parseFloat(pts.shift())
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(x1), svg2cagY(y1)], [svg2cagX(bx), svg2cagY(by)], [svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(bx, by, cx, cy)
-                bx = rf[0]
-                by = rf[1]
-              }
-              break
-            case 'q': // relative quadratic Bézier
-              while (pts.length >= 4) {
-                qx = cx + parseFloat(pts.shift())
-                qy = cy + parseFloat(pts.shift())
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(qx), svg2cagY(qy)], [svg2cagX(qx), svg2cagY(qy)], [svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(qx, qy, cx, cy)
-                qx = rf[0]
-                qy = rf[1]
-              }
-              break
-            case 'Q': // absolute quadratic Bézier
-              while (pts.length >= 4) {
-                qx = parseFloat(pts.shift())
-                qy = parseFloat(pts.shift())
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(qx), svg2cagY(qy)], [svg2cagX(qx), svg2cagY(qy)], [svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(qx, qy, cx, cy)
-                qx = rf[0]
-                qy = rf[1]
-              }
-              break
-            case 't': // relative quadratic Bézier shorthand
-              while (pts.length >= 2) {
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(qx), svg2cagY(qy)], [svg2cagX(qx), svg2cagY(qy)], [cx, cy]])
-                var rf = reflect(qx, qy, cx, cy)
-                qx = rf[0]
-                qy = rf[1]
-              }
-              break
-            case 'T': // absolute quadratic Bézier shorthand
-              while (pts.length >= 2) {
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(qx), svg2cagY(qy)], [svg2cagX(qx), svg2cagY(qy)], [svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(qx, qy, cx, cy)
-                qx = rf[0]
-                qy = rf[1]
-              }
-              break
-            case 's': // relative cubic Bézier shorthand
-              while (pts.length >= 4) {
-                var x1 = bx // reflection of 2nd control point from previous C
-                var y1 = by // reflection of 2nd control point from previous C
-                bx = cx + parseFloat(pts.shift())
-                by = cy + parseFloat(pts.shift())
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(x1), svg2cagY(y1)], [svg2cagX(bx), svg2cagY(by)], [svg2cagX(cx,svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(bx, by, cx, cy)
-                bx = rf[0]
-                by = rf[1]
-              }
-              break
-            case 'S': // absolute cubic Bézier shorthand
-              while (pts.length >= 4) {
-                var x1 = bx // reflection of 2nd control point from previous C
-                var y1 = by // reflection of 2nd control point from previous C
-                bx = parseFloat(pts.shift())
-                by = parseFloat(pts.shift())
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendBezier([[svg2cagX(x1), svg2cagY(y1)], [svg2cagX(bx), svg2cagY(by)], [svg2cagX(cx,svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)]])
-                var rf = reflect(bx, by, cx, cy)
-                bx = rf[0]
-                by = rf[1]
-              }
-              break
-            case 'h': // relative Horzontal line to
-              while (pts.length >= 1) {
-                cx = cx + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'H': // absolute Horzontal line to
-              while (pts.length >= 1) {
-                cx = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'l': // relative line to
-              while (pts.length >= 2) {
-                cx = cx + parseFloat(pts.shift())
-                cy = cy + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'L': // absolute line to
-              while (pts.length >= 2) {
-                cx = parseFloat(pts.shift())
-                cy = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'v': // relative Vertical line to
-              while (pts.length >= 1) {
-                cy = cy + parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'V': // absolute Vertical line to
-              while (pts.length >= 1) {
-                cy = parseFloat(pts.shift())
-                paths[pn] = paths[pn].appendPoint([svg2cagX(cx, svgUnitsPmm), svg2cagY(cy, svgUnitsPmm)])
-              }
-              break
-            case 'z': // close current line
-            case 'Z':
-              code += indent + pn + ' = ' + pn + '.close();\n'
-              code += indent + pn + ' = ' + pn + '.innerToCAG();\n'
-              code += indent + on + ' = ' + on + '.union(' + pn + ');\n'
-              cx = sx; cy = sy // return to the starting point
-              pc = true
-              break
-            default:
-              console.log('Warning: Unknow PATH command [' + co.c + ']')
-              break
-          }
-          // console.log('postion: ['+cx+','+cy+'] after '+co.c);
-        }
-        if (pi > 0) {
-          if (pc === false) {
-            paths[pn] = paths[pn].expandToCAG(r, CSG.defaultResolution2D)
-            pathCag = pathCag.union(paths[pn])
-          }
-        }
-        break
-      default:
-        break
-    }
     if ('fill' in obj) {
     // FIXME when CAG supports color
     //  code += indent+on+' = '+on+'.setColor(['+obj.fill[0]+','+obj.fill[1]+','+obj.fill[2]+']);\n';
     }
     if ('transforms' in obj) {
     // NOTE: SVG specifications require that transforms are applied in the order given.
-    //       But these are applied in the order as required by CSG/CAG
+    // But these are applied in the order as required by CSG/CAG
       var tr = null
       var ts = null
       var tt = null
